Refactor field templates into organized folders
